move settings screen to own folder
